+++ docs/升级日志.md
@@ -2,7 +2,7 @@
 1. 升级 zipkin-server 到 2.19.2 ，并修改项目配置，使得兼容新版本，并完善使用文档
 2. 代码生成器增加 初始化创建服务 的功能 (进行中)
 3. 同步 zuihou-admin-boot 和 zuihou-admin-cloud 项目之间的代码，并部署到新的测试线
-4. 
+4. 优化 jobs项目的配置
 
 # 升级日志 2020年01月09日
 1. 完善数据字典相关接口对接
