+++ zuihou-backend/zuihou-api/pom.xml
@@ -45,7 +45,6 @@
         <dependency>
             <groupId>org.ow2.asm</groupId>
             <artifactId>asm</artifactId>
-            <version>${asm.version}</version>
         </dependency>
 
         <dependency>
