+++ zuihou-backend/zuihou-api/zuihou-authority-api/pom.xml
@@ -19,6 +19,7 @@
         <dependency>
             <groupId>com.github.zuihou</groupId>
             <artifactId>zuihou-authority-entity</artifactId>
+            <version>${project.version}</version>
         </dependency>
     </dependencies>
