+++ zuihou-backend/zuihou-demo/zuihou-demo-server/src/main/java/com/github/zuihou/demo/config/DemoWebConfiguration.java
@@ -1,6 +1,10 @@
 package com.github.zuihou.demo.config;
 
+import com.github.zuihou.authority.api.LogApi;
 import com.github.zuihou.common.config.BaseConfig;
+import com.github.zuihou.log.event.SysLogListener;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 /**
@@ -9,5 +13,11 @@ import org.springframework.context.annotation.Configuration;
  */
 @Configuration
 public class DemoWebConfiguration extends BaseConfig {
+    @Value("${zuihou.mysql.biz-database:zuihou_defaults}")
+    private String database;
 
+    @Bean
+    public SysLogListener sysLogListener(LogApi logApi) {
+        return new SysLogListener(this.database, (log) -> logApi.save(log));
+    }
 }
