+++ zuihou-backend/zuihou-file/pom.xml
@@ -34,6 +34,11 @@
                 <artifactId>zuihou-file-controller</artifactId>
                 <version>${project.version}</version>
             </dependency>
+            <dependency>
+                <groupId>com.github.zuihou</groupId>
+                <artifactId>zuihou-file-entity</artifactId>
+                <version>${project.version}</version>
+            </dependency>
 
         </dependencies>
     </dependencyManagement>
