+++ zuihou-backend/zuihou-file/zuihou-file-server/src/main/java/com/github/zuihou/general/controller/FileGeneralController.java
@@ -31,5 +31,4 @@ public class FileGeneralController {
         return R.success(map);
     }
 
-
 }
